requirements.txt
tritonclient[all]==2.40.0
tensorrt==8.6.1
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
timm==0.9.12
numpy==1.24.3
pillow==10.1.0
opencv-python==4.8.1.78
protobuf==3.20.3
grpcio==1.59.0
pycuda==2022.2.2 